requirements.txt
asgiref==3.5.2
azure-core==1.24.0
azure-storage-blob==12.12.0
bloom-filter==1.3.3
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.2.13
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
